refactor(server): extract message dispatch into server_dispatch.go
Part of #61 M1. Moves handleClientMessage / handleTimeSync /
handleClientState / handleClientGoodbye into a dedicated file so the
M2 Group event-bus work replaces them cleanly.

Mechanical move, no behavior change.
